test(e2e): split exporterImages per exporter
One file per exporter (emailImages / typstPdfImages / reactPdfImages) so
failures and baselines attribute cleanly, with the shared invalid blocks,
export frame and a deduplicated pdf.js rasterize-and-screenshot loop in
exporterTestUtil. Baselines move with their tests; the typst pages are
regenerated with the parity fixes. The typst compile also feeds from the
pdf package's own default font loaders (not a hand-copied list, which is
how a missing bold/italic face once went unnoticed) and declares the
emoji font family so ZWJ sequences shape correctly.
