+++ tests/src/end-to-end/exporters/__screenshots__/reactPdfImages.test.tsx/pdf-export-page-2-chromium-linux.png
